+++ CHANGES.md
@@ -6,5 +6,5 @@
 
 ## 2023-08-xx 0.0.0
 - Initial thing, using `arbitrary-dateparser` and `DateRangeParser` packages
-- Add German variant of `DateRangeParser`
+- Add German variants of `arbitrary-dateparser` and `DateRangeParser`
 - Add `parse_single` method, for parsing a single point in time